Add book instance representation to db structure, implement classes to support CRUD requests for it
Added logic that represents individual copies of each book that the store sells or lends. Each book instance can be sellable, rentable or both (represented by boolean values).
A special enum value represents the availability of the book: 'AVAILABLE', 'RENTED', 'SOLD'. The table also keeps count of rentals, after some set number of rentals the book will be marked as 'TO_BE_DONATED' as it will have earned the store its value. Once it is donated, it will be marked so.
Partially implemented searching logic for easy access (to be finished soon).
Implemented controller and service classes to handle the CRUD operations with initial security