Implement Rental entity and repository, DTOs
Implemented Rental entity class to represent each rental of books in the db.

Each rental keeps the BookInstance entity's id as a foreign key (the copy that is rented), the id of the user that rents/ed it, the timestamp of when the book was rented, the expected return date and the actual return date. The table will also keep track of the status of the rental through an Enum value that will have the following objects: 'IN_PROGRESS', 'RETURNED_IN_TIME', 'RETURNED_LATE', 'OVERDUE'.

Implemented DTO objects for the CRUD requests (to be implemented soon).