Implement Email sender (requires configuration)
Implemented an email sending service that sends a welcome message to newly registered users.
Since this is not a full scale application and hence it has limitations, I configured a personal email to be used for it. So, for privacy, the corresponding configurations have been replaced with placeholders in the application.properties file. Users can configure with their own email credentials.