Implement price, renting, selling logic
Implemented services and controllers to handle the payments for, renting and selling of the books.
Improved and implemented the BookInstance search to include criteria for the new logic.

In my bookstore book instances/copies can be either sellable, rentable or both. The staff member registering the instance inputs that information.
The gem of the bookstore is that we guarantee that the rentable books are sold at a lower price once they are rented at least once (the reduced amount is directly proportional to the number of times the copy was rented) and if it is not sold (or was initially not sellable), it gets donated after some number of times it is rented (determined by the staff member per book instance). This way, we not only contribute to our community, but we give a reason to our users to chose us over other stores. By renting our books they have the guarantee that they will either help another reader buy the book at a more affordable price or help us donate to those in need.